+++ Runtime/RenderPipeline/ScreenSpaceLighting/ForwardGBufferPass.cs
@@ -151,8 +151,13 @@ namespace Illusion.Rendering
                 
                 var drawSettings = CreateDrawingSettings(_shaderTagIdList,
                     ref renderingData, renderingData.cameraData.defaultOpaqueSortFlags);
-                context.DrawRenderers(renderingData.cullResults, ref drawSettings,
-                    ref _filteringSettings, ref _renderStateBlock);
+#if UNITY_2023_1_OR_NEWER
+                var rendererList = default(RendererList);
+                RenderingUtils.CreateRendererListWithRenderStateBlock(context, renderingData, drawSettings, _filteringSettings, _renderStateBlock, ref rendererList);
+                cmd.DrawRendererList(rendererList);
+#else
+                context.DrawRenderers(renderingData.cullResults, ref drawSettings, ref _filteringSettings, ref _renderStateBlock);
+#endif
             }
             context.ExecuteCommandBuffer(cmd);
             CommandBufferPool.Release(cmd);