+++ Runtime/RenderPipeline/Transparency/TransparentDepthOnlyPostPass.cs
@@ -44,8 +44,13 @@ namespace Illusion.Rendering
             context.ExecuteCommandBuffer(cmd);
             var drawSettings = RenderingUtils.CreateDrawingSettings(PostDepthNormalsTagId,
                 ref renderingData, renderingData.cameraData.defaultOpaqueSortFlags);
-            context.DrawRenderers(renderingData.cullResults, ref drawSettings,
-                ref _filteringSettings, ref _renderStateBlock);
+#if UNITY_2023_1_OR_NEWER
+            var rendererList = default(RendererList);
+            RenderingUtils.CreateRendererListWithRenderStateBlock(context, renderingData, drawSettings, _filteringSettings, _renderStateBlock, ref rendererList);
+            cmd.DrawRendererList(rendererList);
+#else
+            context.DrawRenderers(renderingData.cullResults, ref drawSettings, ref _filteringSettings, ref _renderStateBlock);
+#endif
         }
 
         public override void Execute(ScriptableRenderContext context, ref RenderingData renderingData)